Go HTTP Filter: Improve performance
The current HTTP Go filter utilizes a sync.Map to manage the lifecycle of the requests memory between Go and C++. Envoy instances that have a larger number of workers (>=32), the sync.Map causes contention on the underlying lock and reduces performance.

Before this fix, a direct reply benchmark would get about `210,000 req/sec` which is considerably lower than `400,000 req/sec` (Envoy that uses a direct reply within a route with no Go HTTP filter). The same benchmark with this fix included gets `350,000 req/sec`, a 67% increase in performance.
a

Fixes #31916

Signed-off-by: Braden Bassingthwaite <bbassingthwaite@digitalocean.com>